+++ src/main/java/net/sf/jabref/gui/shared/OpenSharedDatabaseDialog.java
@@ -9,6 +9,7 @@ import java.awt.event.KeyEvent;
 import java.io.UnsupportedEncodingException;
 import java.security.GeneralSecurityException;
 import java.sql.SQLException;
+import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 
@@ -30,6 +31,8 @@ import javax.swing.KeyStroke;
 
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefException;
+import net.sf.jabref.JabRefGUI;
+import net.sf.jabref.gui.BasePanel;
 import net.sf.jabref.gui.JabRefFrame;
 import net.sf.jabref.gui.help.HelpAction;
 import net.sf.jabref.logic.help.HelpFile;
@@ -99,6 +102,14 @@ public class OpenSharedDatabaseDialog extends JDialog {
     }
 
     public void openSharedDatabase() {
+
+        if (isSharedDatabaseAlreadyPresent()) {
+            JOptionPane.showMessageDialog(OpenSharedDatabaseDialog.this,
+                    Localization.lang("You are already connected to a database using entered connection details."),
+                    Localization.lang("Warning"), JOptionPane.WARNING_MESSAGE);
+            return;
+        }
+
         setLoadingConnectButtonText(true);
 
         try {
@@ -382,4 +393,18 @@ public class OpenSharedDatabaseDialog extends JDialog {
         }
     }
 
+    /**
+     * Checks whether a database with the given @link {@link DBMSConnectionProperties} is already opened.
+     */
+    private boolean isSharedDatabaseAlreadyPresent() {
+        List<BasePanel> panels = JabRefGUI.getMainFrame().getBasePanelList();
+        for (BasePanel panel : panels) {
+            DBMSConnectionProperties dbmsConnectionProperties = panel.getBibDatabaseContext().getDBMSSynchronizer()
+                    .getDBProcessor().getDBMSConnectionProperties();
+            if (this.connectionProperties.equals(dbmsConnectionProperties)) {
+                return true;
+            }
+        }
+        return false;
+    }
 }